+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Adds [`Range.compareBeginToBegin`][].
 - Adds [`Range.compareEndToEnd`][].
+- Adds [`Range.prototype.compareBegin`][].
 
 - Sets numeric unbounded endpoints to `-Infinity`/`Infinity` if you pass
   `Number` to [`Range.parse`][] as the parse function.
@@ -12,6 +13,7 @@
 
 [`Range.compareBeginToBegin`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.compareBeginToBegin
 [`Range.compareEndToEnd`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.compareEndToEnd
+[`Range.prototype.compareBegin`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.prototype.compareBegin
 
 ## 1.2.0 (Jul 4, 2015)
 - Adds [`Range.prototype.isBounded`][].